fix(quality): decompose AiService into prompt factory, PII redactor, endpoint guard and audit log
AiService was 1265 lines (threshold 1000). It orchestrated AI calls AND owned
four self-contained concerns:

- AiPromptFactory: the six prompt templates, so the wording of what we ask a
  model and the JSON shape we demand back are reviewable as a set.
- AiPiiRedactor: the ONE definition of deterministically-detectable PII plus
  both of its consumers (span reporting and prompt scrubbing), so detection and
  scrubbing can never drift apart on which patterns count.
- AiEndpointGuard: the SSRF decision — CIDR deny-list, cloud/local rules, IPv4
  and IPv6 range arithmetic.
- AiAuditLog: the Algoritmeregister oversight trail, with the write and the read
  resolving the same register/schema config in one place and degrading in one
  place.

AiService keeps orchestration and the single outbound model call. It no longer
needs the SearchesObjects or SuppressesWarnings traits, nor the DI container.

Tests wire the REAL collaborators against the existing mocked boundaries rather
than stubbing them out, so AiServiceAuditLoggingCompletenessTest still proves an
audit entry is actually written and AiServicePiiDetectionTest still exercises
the real pattern set.

phpmd: AiService ExcessiveClassLength 1265 -> cleared; repo total 19 -> 18.
TooManyPublicMethods (12/10) is NOT cleared — see the PR body.
